Fold the back-fill tests into the existing compat file, trim comments
One file for the transformers>=5 config tests instead of two. The two `.ckpt`
tests covered the same path, so they collapse into one that strips every
v5-only field rather than just `scale_decoder_outputs`.
